refactor: moved modules to specific folder
