+++ src/modules/auth/auth.module.ts
@@ -4,7 +4,7 @@ import { JwtModule } from '@nestjs/jwt';
 import { ConfigService, ConfigModule } from '@nestjs/config';
 import { AuthService } from './auth.service';
 import { LocalStrategy } from './local.strategy';
-import { UsersModule } from '~/users/users.module';
+import { UsersModule } from '~/modules/users/users.module';
 import { JwtStrategy } from './jwt.strategy';
 
 @Module({
