+++ src/modules/reset-password/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './apply-reset-password-user.dto';
+export * from './request-reset-password-user.dto';
