+++ src/modules/reset-password/interfaces/index.ts
