+++ src/modules/reset-password/reset-password.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ConfigService } from '@nestjs/config';
 import { ResetPasswordService } from './reset-password.service';
-import { ResetPasswordSchema } from '~/schemas/reset-password.schema';
-import { UsersModule } from '~/users/users.module';
+import { ResetPasswordSchema } from '~/modules/reset-password/schemas/reset-password.schema';
+import { UsersModule } from '~/modules/users/users.module';
 
 @Module({
   imports: [
