+++ src/modules/users/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './create-user.dto';
