+++ src/modules/users/interfaces/create-user-repository.interface.ts
@@ -1,5 +1,5 @@
 import { UserModel } from './user-model.interface';
-import { CreateUserDto } from '~/dto/users/create-user.dto';
+import { CreateUserDto } from '~/modules/users/dto';
 
 export interface CreateUserRepository {
   create(createUserDTO: CreateUserDto): Promise<UserModel>
