+++ src/modules/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { UsersService } from './users.service';
-import { UserSchema } from '~/schemas/user.schema';
+import { UserSchema } from '~/modules/users/schemas/user.schema';
 
 @Module({
   imports: [MongooseModule.forFeature([{ name: 'User', schema: UserSchema }])],
